% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{AnnDataLayerToSeurat}
\alias{AnnDataLayerToSeurat}
\title{Map AnnData layer name to Seurat layer name}
\usage{
AnnDataLayerToSeurat(name)
}
\arguments{
\item{name}{AnnData layer name}
}
\value{
Seurat layer name
}
\description{
Map AnnData layer name to Seurat layer name
}
\keyword{internal}
