% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveH5MU.R
\name{as.h5mu}
\alias{as.h5mu}
\title{Save a Seurat object to h5mu format}
\usage{
as.h5mu(object, filename, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{filename}{Output filename}

\item{...}{Additional arguments passed to \code{\link{writeH5MU}}}
}
\value{
Invisibly returns the filename
}
\description{
Save a Seurat object to h5mu format
}
\seealso{
\code{\link{writeH5MU}}
}
