% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CachedGuessDType}
\alias{CachedGuessDType}
\title{Cached version of GuessDType for known string constants}
\usage{
CachedGuessDType(x, stype = "utf8")
}
\arguments{
\item{x}{An R object to guess the HDF5 dtype for}

\item{stype}{String encoding type (default 'utf8')}
}
\value{
An object of class \code{\link[hdf5r]{H5T}}
}
\description{
Checks an internal cache before calling \code{\link{GuessDType}}.
This avoids repeated H5T allocation for the same AnnData encoding strings.
}
\keyword{internal}
