% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CachedUtf8Type}
\alias{CachedUtf8Type}
\title{Get cached UTF-8 string dtype}
\usage{
CachedUtf8Type()
}
\value{
An \code{\link[hdf5r]{H5T_STRING}} object
}
\description{
Returns a singleton UTF-8 string type, avoiding repeated allocation.
}
\keyword{internal}
