% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialData.R
\name{.add_spatialdata_coords}
\alias{.add_spatialdata_coords}
\title{Add spatial coordinates from SpatialData to Seurat object}
\usage{
.add_spatialdata_coords(
  obj,
  coords,
  radius = 1,
  region_name = "spots",
  cell_names = NULL,
  instance_key = NULL,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Seurat object}

\item{coords}{Coordinate matrix (n x 2) with x, y columns}

\item{radius}{Spot radius}

\item{region_name}{Name of the spatial region}

\item{cell_names}{Cell names in the Seurat object}

\item{instance_key}{Column in obs linking to shapes index}

\item{verbose}{Show progress}
}
\value{
Modified Seurat object
}
\description{
Add spatial coordinates from SpatialData to Seurat object
}
\keyword{internal}
