% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialData.R
\name{.attach_spatialdata_image}
\alias{.attach_spatialdata_image}
\title{Attach an image array to a Seurat object}
\usage{
.attach_spatialdata_image(obj, img_data, img_name = "image", verbose = TRUE)
}
\arguments{
\item{obj}{Seurat object (must already have spatial coordinates)}

\item{img_data}{3D array (H, W, C) normalized to [0, 1]}

\item{img_name}{Name for the image}

\item{verbose}{Show progress}
}
\value{
Modified Seurat object
}
\description{
Attach an image array to a Seurat object
}
\keyword{internal}
