% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.decode_vlen_utf8}
\alias{.decode_vlen_utf8}
\title{Decode vlen-utf8 encoded raw data to character vector}
\usage{
.decode_vlen_utf8(raw_data, n)
}
\arguments{
\item{raw_data}{Raw vector containing vlen-utf8 encoded strings}

\item{n}{Number of strings to decode}
}
\value{
Character vector of length n
}
\description{
Decode vlen-utf8 encoded raw data to character vector
}
\keyword{internal}
