% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.encode_vlen_utf8}
\alias{.encode_vlen_utf8}
\title{Encode character vector to vlen-utf8 raw data}
\usage{
.encode_vlen_utf8(strings)
}
\arguments{
\item{strings}{Character vector to encode}
}
\value{
Raw vector in vlen-utf8 format
}
\description{
Encode character vector to vlen-utf8 raw data
}
\keyword{internal}
