% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamHDF5.R
\name{.h5ad_write_dict_group}
\alias{.h5ad_write_dict_group}
\title{Write an AnnData dict-style empty group}
\usage{
.h5ad_write_dict_group(parent, name)
}
\arguments{
\item{parent}{H5Group or H5File}

\item{name}{Group name}
}
\description{
Write an AnnData dict-style empty group
}
\keyword{internal}
