% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamHDF5.R
\name{.h5seurat_write_scaffold}
\alias{.h5seurat_write_scaffold}
\title{Write required h5seurat root attributes and empty groups}
\usage{
.h5seurat_write_scaffold(h5, assay, n_cells)
}
\arguments{
\item{h5}{H5File handle (mode = "w")}

\item{assay}{Character active assay name}

\item{n_cells}{Integer number of cells (for active.ident)}
}
\description{
Write required h5seurat root attributes and empty groups
}
\keyword{internal}
