% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZarrRemote.R
\name{.is_remote_zarr_url}
\alias{.is_remote_zarr_url}
\title{Detect remote zarr URL}
\usage{
.is_remote_zarr_url(x)
}
\arguments{
\item{x}{A character string (path or URL)}
}
\value{
TRUE if x looks like a remote zarr URL (s3://, gs://, http(s)://)
}
\description{
Detect remote zarr URL
}
\keyword{internal}
