% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamHDF5.R
\name{.modality_to_assay}
\alias{.modality_to_assay}
\title{Map h5mu modality name to Seurat assay name}
\usage{
.modality_to_assay(mod)
}
\arguments{
\item{mod}{Character modality name from h5mu (e.g., "rna", "prot")}
}
\value{
Character assay name (e.g., "RNA", "ADT")
}
\description{
Map h5mu modality name to Seurat assay name
}
\keyword{internal}
