% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.raw_fill_to_r}
\alias{.raw_fill_to_r}
\title{Convert zarr fill_value to R scalar of the right type}
\usage{
.raw_fill_to_r(fill_value, dtype_info)
}
\description{
Convert zarr fill_value to R scalar of the right type
}
\keyword{internal}
