% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_has_write_soma}
\alias{.soma_has_write_soma}
\title{Check whether tiledbsoma::write_soma is available}
\usage{
.soma_has_write_soma()
}
\value{
Logical
}
\description{
Check whether tiledbsoma::write_soma is available
}
\keyword{internal}
