% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_read_extra_measurements}
\alias{.soma_read_extra_measurements}
\title{Read additional measurements as extra assays}
\usage{
.soma_read_extra_measurements(
  exp,
  obj,
  primary,
  n_cells,
  cell_names,
  verbose = TRUE
)
}
\arguments{
\item{exp}{A SOMAExperiment object}

\item{obj}{A Seurat object to add assays to}

\item{primary}{Name of the primary measurement (already loaded)}

\item{n_cells}{Number of cells in the primary measurement}

\item{cell_names}{Cell names}

\item{verbose}{Verbosity}
}
\value{
The Seurat object with extra assays added
}
\description{
Read additional measurements as extra assays
}
\keyword{internal}
