% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_read_obs}
\alias{.soma_read_obs}
\title{Read obs from a SOMA experiment}
\usage{
.soma_read_obs(exp, obs_query = NULL, column_names = NULL)
}
\arguments{
\item{exp}{A SOMAExperiment object}

\item{obs_query}{Optional value filter string}

\item{column_names}{Optional column names to read}
}
\value{
A data.frame with cell metadata, rownames set to cell identifiers
}
\description{
Read obs from a SOMA experiment
}
\keyword{internal}
