% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_read_obsm}
\alias{.soma_read_obsm}
\title{Read obsm embeddings from a SOMA measurement}
\usage{
.soma_read_obsm(ms, obj, measurement, cell_names, verbose = TRUE)
}
\arguments{
\item{ms}{A SOMAMeasurement object}

\item{obj}{A Seurat object to add reductions to}

\item{measurement}{Assay name}

\item{cell_names}{Cell names for rownames}

\item{verbose}{Verbosity}
}
\value{
The Seurat object with embeddings added
}
\description{
Read obsm embeddings from a SOMA measurement
}
\keyword{internal}
