% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_read_obsp}
\alias{.soma_read_obsp}
\title{Read obsp graphs from a SOMA measurement}
\usage{
.soma_read_obsp(ms, obj, measurement, cell_names, verbose = TRUE)
}
\arguments{
\item{ms}{A SOMAMeasurement object}

\item{obj}{A Seurat object to add graphs to}

\item{measurement}{Assay name}

\item{cell_names}{Cell names for row/column names}

\item{verbose}{Verbosity}
}
\value{
The Seurat object with graphs added
}
\description{
Read obsp graphs from a SOMA measurement
}
\keyword{internal}
