% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_read_var}
\alias{.soma_read_var}
\title{Read var from a SOMA measurement}
\usage{
.soma_read_var(ms, var_query = NULL, column_names = NULL)
}
\arguments{
\item{ms}{A SOMAMeasurement object}

\item{var_query}{Optional value filter string}

\item{column_names}{Optional column names to read}
}
\value{
A data.frame with feature metadata, rownames set to feature names
}
\description{
Read var from a SOMA measurement
}
\keyword{internal}
