% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSOMA.R
\name{.soma_read_x}
\alias{.soma_read_x}
\title{Read X matrix from a SOMA measurement}
\usage{
.soma_read_x(ms, n_cells, n_features, cell_names, feature_names)
}
\arguments{
\item{ms}{A SOMAMeasurement object}

\item{n_cells}{Expected number of cells}

\item{n_features}{Expected number of features}

\item{cell_names}{Character vector of cell names}

\item{feature_names}{Character vector of feature names}
}
\value{
A sparse matrix (features x cells)
}
\description{
Read X matrix from a SOMA measurement
}
\keyword{internal}
