% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamHDF5.R
\name{.stream_dense_matrix}
\alias{.stream_dense_matrix}
\title{Stream a dense matrix from one HDF5 file to another}
\usage{
.stream_dense_matrix(
  src_dataset,
  dst_parent,
  dst_name,
  gzip,
  transpose = FALSE
)
}
\arguments{
\item{src_dataset}{H5D source dataset}

\item{dst_parent}{H5Group to write into}

\item{dst_name}{Name for new dataset}

\item{gzip}{Integer gzip compression level}

\item{transpose}{Logical, whether to transpose}
}
\description{
Reads a dense dataset and writes it with gzip compression.
}
\keyword{internal}
