% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_chunk_path_v2}
\alias{.zarr_chunk_path_v2}
\title{Build chunk file path for zarr v2}
\usage{
.zarr_chunk_path_v2(array_dir, chunk_coords)
}
\arguments{
\item{array_dir}{Array directory path}

\item{chunk_coords}{Integer vector of chunk coordinates (0-based)}
}
\value{
File path to chunk
}
\description{
Build chunk file path for zarr v2
}
\keyword{internal}
