% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_compress}
\alias{.zarr_compress}
\title{Compress chunk data}
\usage{
.zarr_compress(raw_data, compressor)
}
\arguments{
\item{raw_data}{Raw vector to compress}

\item{compressor}{Compressor specification list}
}
\value{
Compressed raw vector
}
\description{
Compress chunk data
}
\keyword{internal}
