% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_decompress}
\alias{.zarr_decompress}
\title{Decompress chunk data}
\usage{
.zarr_decompress(raw_data, compressor)
}
\arguments{
\item{raw_data}{Raw vector of compressed data}

\item{compressor}{Compressor specification list from .zarray metadata}
}
\value{
Decompressed raw vector
}
\description{
Decompress chunk data
}
\keyword{internal}
