% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_extract_compressor_v3}
\alias{.zarr_extract_compressor_v3}
\title{Extract compressor spec from zarr v3 codecs array}
\usage{
.zarr_extract_compressor_v3(codecs)
}
\description{
Extract compressor spec from zarr v3 codecs array
}
\keyword{internal}
