% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_fetch_chunk_strings}
\alias{.zarr_fetch_chunk_strings}
\title{Fetch and decode one (outer) chunk of a zarr string array}
\usage{
.zarr_fetch_chunk_strings(store, chunk_key, meta, n)
}
\value{
character vector of length \code{n}
}
\description{
Fetch and decode one (outer) chunk of a zarr string array
}
\keyword{internal}
