% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_fill_bytes}
\alias{.zarr_fill_bytes}
\title{Byte pattern of the fill value for one element of a zarr array}
\usage{
.zarr_fill_bytes(meta)
}
\description{
Byte pattern of the fill value for one element of a zarr array
}
\keyword{internal}
