% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadZarr.R
\name{.zarr_get_n_obs}
\alias{.zarr_get_n_obs}
\title{Get number of observations from zarr store}
\usage{
.zarr_get_n_obs(store_path)
}
\description{
Get number of observations from zarr store
}
\keyword{internal}
