% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadZarr.R
\name{.zarr_get_n_vars}
\alias{.zarr_get_n_vars}
\title{Get number of variables from zarr store}
\usage{
.zarr_get_n_vars(store_path)
}
\description{
Get number of variables from zarr store
}
\keyword{internal}
