% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_list_children}
\alias{.zarr_list_children}
\title{List children of a zarr group}
\usage{
.zarr_list_children(store_path, rel_path = "")
}
\arguments{
\item{store_path}{Path to zarr store}

\item{rel_path}{Relative path to group within store}
}
\value{
Character vector of child names
}
\description{
List children of a zarr group
}
\keyword{internal}
