% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_node_type}
\alias{.zarr_node_type}
\title{Determine zarr node type}
\usage{
.zarr_node_type(store_path, rel_path)
}
\arguments{
\item{store_path}{Path to zarr store}

\item{rel_path}{Relative path within store}
}
\value{
"group", "array", or "missing"
}
\description{
Determine zarr node type
}
\keyword{internal}
