% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_parse_dtype}
\alias{.zarr_parse_dtype}
\title{Parse zarr v2 dtype string to R type info}
\usage{
.zarr_parse_dtype(dtype)
}
\arguments{
\item{dtype}{Zarr dtype string (e.g., "<f8", "<i4", "|O")}
}
\value{
List with r_type, size, endian, is_object
}
\description{
Parse zarr v2 dtype string to R type info
}
\keyword{internal}
