% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_parse_dtype_v3}
\alias{.zarr_parse_dtype_v3}
\title{Parse zarr v3 data_type string into dtype_info struct}
\usage{
.zarr_parse_dtype_v3(data_type, codecs = NULL)
}
\description{
Parse zarr v3 data_type string into dtype_info struct
}
\keyword{internal}
