% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadZarr.R
\name{.zarr_read_anndata_index}
\alias{.zarr_read_anndata_index}
\title{Read AnnData index from obs or var group}
\usage{
.zarr_read_anndata_index(store_path, group)
}
\arguments{
\item{store_path}{Path to zarr store}

\item{group}{"obs" or "var"}
}
\value{
Character vector of index values, or NULL
}
\description{
Read AnnData index from obs or var group
}
\keyword{internal}
