% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialData.R
\name{.zarr_read_anndata_index_from_attrs}
\alias{.zarr_read_anndata_index_from_attrs}
\title{Read an anndata index given pre-read attrs}
\usage{
.zarr_read_anndata_index_from_attrs(store_path, attrs)
}
\arguments{
\item{store_path}{Path to zarr group}

\item{attrs}{Pre-parsed .zattrs list}
}
\value{
Character vector of index values
}
\description{
Read an anndata index given pre-read attrs
}
\keyword{internal}
