% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_read_chunked}
\alias{.zarr_read_chunked}
\title{Read multi-chunk zarr v2 numeric array}
\usage{
.zarr_read_chunked(
  store_or_path,
  array_rel,
  meta,
  dtype_info,
  shape,
  chunks,
  n_chunks_per_dim,
  compressor
)
}
\description{
Read multi-chunk zarr v2 numeric array
}
\keyword{internal}
