% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_read_json}
\alias{.zarr_read_json}
\title{Read and parse a JSON file}
\usage{
.zarr_read_json(path)
}
\arguments{
\item{path}{Path to JSON file}
}
\value{
Parsed list
}
\description{
Read and parse a JSON file
}
\keyword{internal}
