% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_store_version}
\alias{.zarr_store_version}
\title{Detect zarr store format version}
\usage{
.zarr_store_version(store_path)
}
\arguments{
\item{store_path}{Path to zarr store directory}
}
\value{
2L or 3L
}
\description{
Detect zarr store format version
}
\keyword{internal}
