% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZarrRemote.R
\name{.zarr_translate_url}
\alias{.zarr_translate_url}
\title{Translate cloud URL scheme to HTTPS base URL}
\usage{
.zarr_translate_url(url)
}
\arguments{
\item{url}{A character URL}
}
\value{
A list with `base` (https URL, trailing slash stripped) and
  `list_url` (for ListObjectsV2 enumeration on S3/GCS) and `kind`
  ("s3", "gs", "http").
}
\description{
s3://bucket/key -> https://bucket.s3.amazonaws.com/key (virtual-hosted style,
anonymous public buckets only).
gs://bucket/key -> https://storage.googleapis.com/bucket/key.
http(s):// is returned as-is.
}
\keyword{internal}
