% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_write_attrs}
\alias{.zarr_write_attrs}
\title{Write zarr attributes to .zattrs file}
\usage{
.zarr_write_attrs(dir, attrs)
}
\arguments{
\item{dir}{Directory containing the zarr node}

\item{attrs}{Named list of attributes}
}
\description{
Write zarr attributes to .zattrs file
}
\keyword{internal}
