% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_write_strings}
\alias{.zarr_write_strings}
\title{Write a zarr v2 vlen-utf8 string array}
\usage{
.zarr_write_strings(dir, strings, compressor = NULL, attrs = NULL)
}
\arguments{
\item{dir}{Directory for the array}

\item{strings}{Character vector}

\item{compressor}{Compressor spec list (default gzip level 4)}

\item{attrs}{Optional attributes for .zattrs}
}
\description{
Write a zarr v2 vlen-utf8 string array
}
\keyword{internal}
