% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{EncodeCategorical}
\alias{EncodeCategorical}
\title{Encode R factor as AnnData categorical}
\usage{
EncodeCategorical(x)
}
\arguments{
\item{x}{A factor (or coercible to factor)}
}
\value{
A list with \code{codes} (0-based integer) and \code{categories}
  (character vector of levels)
}
\description{
Encode R factor as AnnData categorical
}
\keyword{internal}
