% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialFOV.R
\name{ExtractFOVCentroids}
\alias{ExtractFOVCentroids}
\title{Extract centroid coordinates from an FOV object}
\usage{
ExtractFOVCentroids(fov_obj)
}
\arguments{
\item{fov_obj}{FOV (or nested Centroids/Segmentation) object}
}
\value{
numeric matrix (n_cells x 2) with colnames c("x","y"), or NULL
}
\description{
Extract centroid coordinates from an FOV object
}
\keyword{internal}
