% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FileType}
\alias{FileType}
\title{Convert an HDF5 compound dataset to a group (legacy version)}
\usage{
FileType(file)
}
\arguments{
\item{file}{Name of file}

\item{src}{An HDF5 dataset}

\item{dest}{An HDF5 file or group}

\item{dst.name}{Name of group in \code{dest}
Determine a filetype based on its extension}
}
\value{
The extension, all lowercase
}
\description{
Convert an HDF5 compound dataset to a group (legacy version)
}
\examples{
\donttest{
scConvert:::FileType('pbmc3k.h5Seurat')
scConvert:::FileType('h5ad')
}

}
\keyword{internal}
