% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ScalarSpace}
\alias{ScalarSpace}
\title{Get cached scalar HDF5 dataspace}
\usage{
ScalarSpace()
}
\value{
An \code{\link[hdf5r]{H5S}} object of type "scalar"
}
\description{
Returns a singleton H5S scalar dataspace, avoiding repeated allocation.
}
\keyword{internal}
