% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transpose.R
\name{scTranspose}
\alias{scTranspose}
\alias{scTranspose.dgCMatrix}
\alias{scTranspose.H5D}
\alias{scTranspose.H5Group}
\title{scTranspose a matrix}
\usage{
scTranspose(x, ...)

\method{scTranspose}{dgCMatrix}(x, ...)

\method{scTranspose}{H5D}(
  x,
  dest = GetParent(x = x),
  dname = paste0("t_", basename(path = x$get_obj_name())),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)

\method{scTranspose}{H5Group}(
  x,
  dest = GetParent(x = x),
  dname = paste0("t_", basename(path = x$get_obj_name())),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{A matrix to transpose}

\item{...}{Arguments passed to other methods}

\item{dest}{...}

\item{dname}{...}

\item{overwrite}{...}

\item{verbose}{Show progress updates}
}
\value{
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} method: returns a
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} with the data of \code{x} transposed

\code{\link[hdf5r]{H5D}} and \code{\link[hdf5r]{H5Group}} methods:
Invisibly returns \code{NULL}
}
\description{
scTranspose a matrix
}
