% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{SeuratLayerToAnnData}
\alias{SeuratLayerToAnnData}
\title{Map Seurat layer name to AnnData layer name}
\usage{
SeuratLayerToAnnData(name)
}
\arguments{
\item{name}{Seurat layer name}
}
\value{
AnnData layer name
}
\description{
Map Seurat layer name to AnnData layer name
}
\keyword{internal}
