% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{DecodeCategorical}
\alias{DecodeCategorical}
\title{Decode AnnData categorical encoding to R factor}
\usage{
DecodeCategorical(codes, categories)
}
\arguments{
\item{codes}{Integer vector of 0-based category codes (-1 = NA)}

\item{categories}{Character vector of category labels}
}
\value{
A factor vector
}
\description{
Decode AnnData categorical encoding to R factor
}
\keyword{internal}
