% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnDataEncoding.R
\name{.zarr_read_numeric}
\alias{.zarr_read_numeric}
\title{Read a zarr v2 numeric array from disk}
\usage{
.zarr_read_numeric(store_path, rel_path)
}
\arguments{
\item{store_path}{Path to zarr store}

\item{rel_path}{Relative path to array within store}
}
\value{
Numeric vector or matrix
}
\description{
Read a zarr v2 numeric array from disk
}
\keyword{internal}
